requirements.txt
flask
pillow
numpy
torch
torchvision